requirements.txt
# Core LangChain
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.0.50

# LLM Providers & Frameworks
openai>=1.0.0
pydantic-ai>=0.1.0

# Server & Utilities
fastapi>=0.100.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0